# Optimize Monomial, Constant to run in constant time,