--- a/polynomial/core.py
+++ b/polynomial/core.py
@@ -34,6 +34,7 @@ def extract_polynomial(method):
 
     If casting is not possible or not appropriate, raise a ValueError.
     """
+
     def decorated(self, other):
         if isinstance(other, Polynomial):
             return method(self, other)
@@ -42,7 +43,7 @@ def extract_polynomial(method):
 
         raise ValueError(
             "{0}.{1} requires a Polynomial or number, got {2}."
-            .format(
+                .format(
                 self.__class__.__name__,
                 method.__name__,
                 type(other).__name__
@@ -90,7 +91,7 @@ def _degree(vec, tuples=True):
 def _mul(lhs, rhs):
     """Return lhs * rhs."""
     if not lhs or not rhs:
-        return
+        return [(0, 0)]
     deg = _degree(lhs) + _degree(rhs) + 1
 
     res = [0] * deg
@@ -169,6 +170,7 @@ class Polynomial:
 
                 raise TypeError("{} cannot be a monomial.".
                                 format(monomial))
+
             self.terms = [monomial_to_tuple(monomial) for monomial in iterable]
         else:
             self._vector = _trim(list(iterable)[::-1])
@@ -229,7 +231,7 @@ class Polynomial:
 
         return Polynomial(
             [c * x for c, x
-                in zip(self, reversed(factors))]
+             in zip(self, reversed(factors))]
         )
 
     @property
@@ -366,7 +368,7 @@ degree {0} of a {1}-degree polynomial".format(degree, self.degree))
         # eg. -         5       x^     2
         s_d = self.degree
         terms = ["{0}{1}{2}{3}".
-                 format(*components(ak, k, k == s_d))
+                     format(*components(ak, k, k == s_d))
                  for ak, k in self.terms]
 
         return " ".join(terms)
@@ -619,7 +621,7 @@ degree {0} of a {1}-degree polynomial".format(degree, self.degree))
         raise ValueError(
             "Can not check {0} for membership. A two-tuple, list of "
             "two-tuples, a set, or a Polynomial are required."
-            .format(type(item).__name__)
+                .format(type(item).__name__)
         )
 
     def terms_are_valid(self, terms):
@@ -639,11 +641,13 @@ degree {0} of a {1}-degree polynomial".format(degree, self.degree))
 
 def setvalue_decorator(error, _terms_are_valid, _fn):
     """Decorate __setattr__, checking if self._vector is still valid."""
+
     def method(self, name, new_value):
         _fn(self, name, new_value)
         if (name == '_vector' and
-                not _terms_are_valid(self, _to_terms(self._vector))):
+            not _terms_are_valid(self, _to_terms(self._vector))):
             raise error
+
     return method
 
 
@@ -654,7 +658,7 @@ class FixedDegreePolynomial(Polynomial):
         """Init a subclass of self."""
         deg = kwargs["valid_degrees"]
         if not isinstance(deg, tuple):
-            deg = (deg, )
+            deg = (deg,)
 
         cls.valid_degrees = deg
 
@@ -731,8 +735,6 @@ class Monomial(FixedTermPolynomial, valid_term_counts=(0, 1)):
         """
         if self._coeff == 0:
             return [(0, 0)]
-        if self._degree == -inf:
-            return [(0, 0)]
         return [(self._coeff, self._degree)]
 
     @terms.setter
@@ -752,21 +754,24 @@ class Monomial(FixedTermPolynomial, valid_term_counts=(0, 1)):
                 err_msg = "terms has more than one non-zero term."
                 curr_coeff, curr_deg = terms[0]
                 termx = []
+
                 for coeff, deg in terms[1:]:
                     if curr_deg == deg:
                         curr_coeff += coeff
-                    elif curr_coeff != 0:
-                        if termx:
-                            raise TermError(err_msg)
-                        termx.append((curr_coeff, curr_deg))
-                        curr_coeff, curr_deg = coeff, deg
+                    else:
+                        if curr_coeff != 0:
+                            if termx:
+                                raise TermError(err_msg)
+                            termx.append((curr_coeff, curr_deg))
+                        curr_coeff = coeff
+                        curr_deg = deg
                 if termx:
                     if curr_coeff:
                         raise TermError(err_msg)
                     self._coeff, self._degree = termx[0]
                 else:
-                    self.coeff = curr_coeff
-                    self.degree = curr_deg
+                    self._coeff = curr_coeff
+                    self._degree = curr_deg
 
     @property
     def _vector(self):
@@ -952,11 +957,13 @@ degree {0} of a {1}-degree monomial".format(degree, self.degree))
     def __setitem__(self, degree, new_value):
         """Set the coefficient of the term with the given degree."""
         if isinstance(degree, slice):
-            self._vector[degree] = new_value
+            _vector = self._vector
+            _vector[degree] = new_value
+            self._vector = _vector
         elif degree == self.degree:
             self.coefficient = new_value
         else:
-            raise IndexError("Can not set more than 1 term on Monomial.")
+            raise TermError("Can not set more than 1 term on Monomial.")
 
 
 class Constant(FixedDegreePolynomial, Monomial, valid_degrees=(0, -inf)):
@@ -976,6 +983,16 @@ class Constant(FixedDegreePolynomial, Monomial, valid_degrees=(0, -inf)):
         if other == self.const:
             return True
         return super().__eq__(other)
+
+    @property
+    def degree(self):
+        """Return self.degree."""
+        return 0 if self._coeff else -inf
+
+    @degree.setter
+    def degree(self, degree):
+        """Set self.degree"""
+        raise DegreeError("Can't change the degree of Constant")
 
     @property
     def const(self):