This adds a _degree, _coeff value to Monomial and subclasses - this avoids constructing a list for single-term monomials, and offers significant speed ups. I've included a number of timings below (100k runs each, run with multiple sets of degrees). Note that for things like initialization, the speed-ups affect any operation making use of Monomials and subclasses: for instance, the QuadraticTrinomial functions below:

```
old real_factors: 0.2728s
new real_factors: 0.1846s
old complex_factors: 0.2666s
new complex_factors: 0.2041s
```

```
   Monomial __init__ degree  1: 0.1428s
   Monomial __init__ degree  6: 0.1561s
   Monomial __init__ degree 11: 0.1252s
   Monomial __init__ degree 16: 0.1300s
   Monomial __init__ degree 21: 0.1377s
OldMonomial __init__ degree  1: 0.3627s
OldMonomial __init__ degree  6: 0.4092s
OldMonomial __init__ degree 11: 0.5138s
OldMonomial __init__ degree 16: 0.4832s
OldMonomial __init__ degree 21: 0.5136s
   Monomial a.a degree  1: 0.1649s
   Monomial a.a degree  6: 0.1508s
   Monomial a.a degree 11: 0.1731s
   Monomial a.a degree 16: 0.1733s
   Monomial a.a degree 21: 0.1421s
OldMonomial a.a degree  1: 0.6473s
OldMonomial a.a degree  6: 0.9247s
OldMonomial a.a degree 11: 0.8272s
OldMonomial a.a degree 16: 0.8342s
OldMonomial a.a degree 21: 0.9063s
   Monomial a + a degree  1: 1.7398s
   Monomial a + a degree  6: 1.8574s
   Monomial a + a degree 11: 1.8001s
   Monomial a + a degree 16: 1.8330s
   Monomial a + a degree 21: 1.8199s
OldMonomial a + a degree  1: 2.6936s
OldMonomial a + a degree  6: 2.6635s
OldMonomial a + a degree 11: 3.0994s
OldMonomial a + a degree 16: 4.5472s
OldMonomial a + a degree 21: 3.6948s
   Monomial a * a degree  1: 0.6333s
   Monomial a * a degree  6: 0.6229s
   Monomial a * a degree 11: 0.6699s
   Monomial a * a degree 16: 0.7277s
   Monomial a * a degree 21: 0.6959s
OldMonomial a * a degree  1: 3.9711s
OldMonomial a * a degree  6: 4.8216s
OldMonomial a * a degree 11: 5.4049s
OldMonomial a * a degree 16: 5.8327s
OldMonomial a * a degree 21: 6.3129s
   Monomial repr(a) degree  1: 0.0916s
   Monomial repr(a) degree  6: 0.0916s
   Monomial repr(a) degree 11: 0.1010s
   Monomial repr(a) degree 16: 0.0937s
   Monomial repr(a) degree 21: 0.0913s
OldMonomial repr(a) degree  1: 1.0931s
OldMonomial repr(a) degree  6: 1.1074s
OldMonomial repr(a) degree 11: 1.2376s
OldMonomial repr(a) degree 16: 1.2969s
OldMonomial repr(a) degree 21: 1.3285s
   Monomial a.terms degree  1: 0.0368s
   Monomial a.terms degree  6: 0.0378s
   Monomial a.terms degree 11: 0.0401s
   Monomial a.terms degree 16: 0.0389s
   Monomial a.terms degree 21: 0.0370s
OldMonomial a.terms degree  1: 0.3945s
OldMonomial a.terms degree  6: 0.5054s
OldMonomial a.terms degree 11: 0.4568s
OldMonomial a.terms degree 16: 0.5439s
OldMonomial a.terms degree 21: 0.6453s
   Monomial a.zero_instance(): 0.1282s
OldMonomial a.zero_instance(): 0.1274s
   Monomial a.terms multiple non-zero accept: 0.4089s
OldMonomial a.terms multiple non-zero accept: 0.6174s
   Monomial set a.terms one term accept: 0.1702s
OldMonomial set a.terms one term accept: 0.5346s
   Monomial a.terms multiple degrees accept: 0.4176s
OldMonomial a.terms multiple degrees accept: 0.5418s
   Monomial a.terms extremely large number accept: 0.4837s
OldMonomial a.terms extremely large number accept: 1.8962s
```