@pep8speaks 2020-07-07
Hello @philippeitis! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`polynomial/core.py`](https://github.com/allexks/py-polynomial/blob/838247ac116cddfa3561515258f628a29dbf2c92/polynomial/core.py):

> [Line 750:80](https://github.com/allexks/py-polynomial/blob/838247ac116cddfa3561515258f628a29dbf2c92/polynomial/core.py#L750): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (88 > 79 characters)
> [Line 762:80](https://github.com/allexks/py-polynomial/blob/838247ac116cddfa3561515258f628a29dbf2c92/polynomial/core.py#L762): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (85 > 79 characters)
> [Line 768:80](https://github.com/allexks/py-polynomial/blob/838247ac116cddfa3561515258f628a29dbf2c92/polynomial/core.py#L768): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (81 > 79 characters)
> [Line 833:17](https://github.com/allexks/py-polynomial/blob/838247ac116cddfa3561515258f628a29dbf2c92/polynomial/core.py#L833): [E128](https://duckduckgo.com/?q=pep8%20E128) continuation line under-indented for visual indent